CLOUDP-301224: Refactor
